fix: dynamic keys from api

+++ src/components/Delete.jsx
@@ -1,16 +1,36 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import Getemp from "./Getemp";
 import FormComp from "./Form";
 
 const DeleteComp = () => {
   const [empId, setEmpId] = useState("");
-  const [fname, setFname] = useState("");
-  const [lname, setLname] = useState("");
-  const [email, setEmail] = useState("");
-  const [dob, setDob] = useState("");
-  const [gender, setGender] = useState("");
   const [loading, setLoading] = useState(false);
   const [norecord, setNoRecord] = useState(false);
+
+  const [baseArr, setBaseArr] = useState([]);
+  const [baseObj, setBaseObj] = useState({});
+
+  useEffect(() => {
+    fetch("https://609a5ad40f5a13001721aac8.mockapi.io/employee_db/1")
+      .then((res) => res.json())
+      .then((res) => {
+        const keysArr = Object.keys(res);
+        const keysFilteredArr = keysArr.filter((item) => item !== "id");
+
+        let newobj = {};
+        for (let i = 0; i < keysFilteredArr.length; i++) {
+          newobj[keysFilteredArr[i]] = "";
+        }
+        setBaseArr(keysFilteredArr);
+        setBaseObj(newobj);
+      })
+      .catch((err) => {
+        console.log(err);
+      })
+      .finally(() => {
+        //setLoading(false);
+      });
+  }, []);
 
   const readHandler = () => {
     if (empId) {
@@ -19,19 +39,12 @@ const DeleteComp = () => {
         .then((res) => res.json())
         .then((res) => {
           if (res !== "Not found") {
-            const { fname, lname, email, dob, gender } = res;
-            setFname(fname);
-            setLname(lname);
-            setEmail(email);
-            setDob(dob);
-            setGender(gender);
+            setBaseObj(res);
             setNoRecord(false);
           } else {
-            setFname("");
-            setLname("");
-            setEmail("");
-            setDob("");
-            setGender("");
+            const temp = { ...baseObj };
+            Object.keys(temp).forEach((key) => (temp[key] = ""));
+            setBaseObj(temp);
             setNoRecord(true);
           }
         })
@@ -47,6 +60,10 @@ const DeleteComp = () => {
   };
 
   const deleteHandler = () => {
+    if (empId == 1) {
+      alert("Base user cant delete");
+      return false;
+    }
     fetch("https://609a5ad40f5a13001721aac8.mockapi.io/employee_db/" + empId, {
       method: "DELETE",
     })
@@ -56,6 +73,9 @@ const DeleteComp = () => {
       .then((json) => {
         if (json) {
           alert("User deleted successfully");
+          const temp = { ...baseObj };
+          Object.keys(temp).forEach((key) => (temp[key] = ""));
+          setBaseObj(temp);
         }
       });
   };
@@ -68,20 +88,13 @@ const DeleteComp = () => {
         readHandler={readHandler}
       />
       <FormComp
-        fname={fname}
-        lname={lname}
-        email={email}
-        dob={dob}
-        gender={gender}
-        setFname={setFname}
-        setLname={setLname}
-        setEmail={setEmail}
-        setDob={setDob}
-        setGender={setGender}
         clicktitle="Delete"
         clickHandler={deleteHandler}
         loading={loading}
         norecord={norecord}
+        baseArr={baseArr}
+        baseObj={baseObj}
+        setBaseObj={setBaseObj}
       />
     </div>
   );
